requirements.txt
absl-py==0.2.0
alabaster==0.7.10
alembic==0.9.9
apturl==0.5.2
astor==0.6.2
astroid==1.6.3
async-generator==1.9
Babel==2.5.3
backcall==0.1.0
beautifulsoup4==4.4.1
bert-serving-client==1.8.1
bert-serving-server==1.8.1
bleach==1.5.0
blinker==1.3
boto==2.49.0
boto3==1.9.25
botocore==1.12.25
bottle==0.12.13
Brlapi==0.6.4
bz2file==0.98
certifi==2018.4.16
chardet==3.0.4
checkbox-support==0.22
Click==7.0
cloudpickle==0.5.2
command-not-found==0.3
cryptography==1.2.3
cycler==0.10.0
cymem==1.31.2
Cython==0.29.3
cytoolz==0.8.2
decorator==4.3.0
defer==1.0.6
dill==0.2.7.1
docutils==0.14
en-core-web-sm==2.0.0
entrypoints==0.2.3
feedparser==5.1.3
Flask==1.0.2
ftfy==4.4.3
gast==0.2.0
gensim==3.6.0
GPUtil==1.4.0
grpcio==1.11.0
guacamole==0.9.2
h5py==2.8.0
html5lib==0.9999999
httplib2==0.9.1
idna==2.7
imagesize==1.0.0
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==1.1.0
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
joblib==0.12.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterhub==0.9.0
Keras==2.1.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
language-selector==0.1
lazy-object-proxy==1.3.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
msgpack-numpy==0.4.1
msgpack-python==0.5.4
murmurhash==0.28.0
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.3
notebook==5.5.0
numpy==1.14.3
numpydoc==0.8.0
oauthlib==1.0.3
onboard==1.2.0
packaging==17.1
padme==1.1.1
pamela==0.3.0
pandas==0.22.0
pandocfilters==1.4.2
parso==0.2.0
pathlib==1.0.1
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.2.0
pkg-resources==0.0.0
plac==0.9.6
plainbox==0.25
POT==0.5.1
preshed==1.0.0
prettytable==0.7.2
prometheus-client==0.2.0
prompt-toolkit==1.0.15
protobuf==3.6.1
psutil==5.4.5
ptyprocess==0.5.2
pyasn1==0.1.9
pycodestyle==2.4.0
pycups==1.9.73
pycurl==7.43.0
pydot==1.2.4
pyflakes==1.6.0
Pygments==2.2.0
pygobject==3.20.0
PyJWT==1.3.0
pylint==1.8.4
PyOpenGL==3.1.0
pyparsing==2.2.0
PyQt5==5.9.2
python-apt==1.1.0b1+ubuntu0.16.4.1
python-dateutil==2.7.2
python-debian==0.1.27
python-editor==1.0.3
python-oauth2==1.1.0
python-systemd==231
pytz==2018.4
pyxdg==0.25
PyYAML==3.12
pyzmq==17.1.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.2
regex==2017.4.5
reportlab==3.3.0
requests==2.19.1
rope==0.10.7
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.0.1
screen-resolution-extra==0.0.0
Send2Trash==1.5.0
sessioninstaller==0.0.0
simplegeneric==0.8.1
sip==4.19.8
six==1.11.0
sklearn==0.0
smart-open==1.7.1
snowballstemmer==1.2.1
spacy==2.0.9
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
SQLAlchemy==1.2.7
ssh-import-id==5.5
stanfordcorenlp==3.9.1.1
stats==0.1.2a0
sudospawner==0.5.2
system-service==0.3
tensorboard==1.11.0
tensorflow-gpu==1.11.0
tensorflow-hub==0.1.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
thinc==6.10.2
toolz==0.9.0
torch==0.4.0
torchvision==0.2.1
tornado==5.0.2
tqdm==4.23.1
traitlets==4.3.2
transliterate==1.10.2
ubuntu-drivers-common==0.0.0
ujson==1.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.23
usb-creator==0.3.0
virtualenv==15.0.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wikiextractor==2.69
wrapt==1.10.11
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
xmltodict==0.11.0
